refactor(openomni): split extension manager helpers
Move audit text utilities and extension manifest audit helpers out of the extension manager while preserving existing manager type re-exports and lifecycle behavior.
